Updates notebook for new TAZ data
Updates the notebook to utilize the new TAZ dataset received from UDOT, which reflects data as of July 2025.
This change ensures that K-12 enrollment data is correctly joined and processed with the latest TAZ information.

Adds new TAZ data files and associated metadata. Updates file paths and variable names in the notebooks to reflect the new dataset.
This ensures the notebooks correctly process K-12 enrollment data with the updated TAZ information. Includes a note about the source of the TAZ file.

Removes outdated result files and index files.